Refactor delete test plus coverage improvement.
